requirements.txt
ultralytics
opencv-python-headless
numpy
tqdm
matplotlib
pyyaml
pillow
